pyproject.toml
[project]
name = "square-mcp"
version = "0.1.0"
description = "Square API Model Context Protocol Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.2.0",
    "squareup>=33.1.0",
]

[project.scripts]
square-mcp = "square_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
